prepare for clip system
fixed memory leak when switching screens:
turns out menu texture was not getting destroyed

also all text objects need to be destroyed as well, the leakage was the opened font files never getting closed.

added a bunch of functionality for the second level, it is incomplete as some rules need to be changed.